app/backend/requirements.txt
pyautogen>=0.2.0
python-dotenv>=1.0.0
openai>=1.0.0
Flask>=2.0.0
requests>=2.20.0